pyproject.toml
[tool.poetry]
name = "scheduling_system"
version = "0.1.0"
description = ""
authors = ["alaa hasan"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.3"
graphene-django = "^2.15.0"
django-graphql-jwt = "=0.3.0"
django-graphql-auth = "^0.3.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
